% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dna.R
\name{plot_dna}
\alias{plot_dna}
\title{Plot DNA sequence}
\usage{
plot_dna(seq)
}
\arguments{
\item{seq}{A character string representing the DNA sequence.}
}
\value{
A \code{ggplot2} object representing the DNA sequence plot.
}
\description{
Creates a plot of a DNA sequence, with each base represented by a colored rectangle.
}
\examples{

seq <- "ATCGATCG"
plot_dna(seq)

}
